requirements.txt
Django==1.11.8
django-crispy-forms==1.6.1
django-formtools==2.0
django-reversion==2.0.10
djangorestframework==3.6.3
future==0.16.0
httplib2==0.9.2
PyMySQL==0.7.11
pytz==2017.2
six==1.10.0
#xadmin==0.6.1
django-import-export==0.5.1
unipath==1.1
django-rest-swagger==2.1.0
pillow==3.4.2
django-extensions==1.8.1
django-filter==1.0.4
boto3==1.4.7
oss2==2.3.3
#django_sendmail_backend==0.1.2
django-registration==2.2
flake8==3.4.1
#pip install https://github.com/CoinLQ/xadmin/archive/master.zip
#pip install https://github.com/CoinLQ/django-fagungis/archive/master.zip
#pip install https://github.com/CoinLQ/db_file_storage/archive/master.zip
#Fabric3==1.13.1.post1
django-storages==1.6.5
#django-db-file-storage==0.4.2
dotmap==1.2.20
numpy==1.13.1
django-redis==4.6
codecov==2.0.9

django-template-debug
django-material
celery==4.1.0
django-celery-beat==1.1.0
django-celery-results==1.0.1
